Rearrange to facilitate better Layout class
